twitter-bot/requirements.txt
fastapi
tweepy
openai
transformers
uvicorn
python-dotenv